Move methods to util.rb
